test(stencils): split stencil-actions.test.ts by concern
